requirements.txt
discord.py
python-dotenv
discord-py-slash-command
youtube-dl
PyNaCl
ffmpeg
progressbar
bs4
requests
datetime